refactor: remove duplicate code (#686)
